Final refactoring finished
