+++ src/components/Buttons/AddButton.js
@@ -1,4 +1,5 @@
 import React from 'react';
+
 import styled  from 'styled-components';
 import  plus from '../../assets/icon-plus.svg';
