+++ src/components/Buttons/ButtonCancel.js
@@ -1,4 +1,5 @@
 import React from 'react';
+
 import { ButtonCancelStyle } from './ButtonsGlobalStyle';
 
 const ButtonCancel = ({ onCancelDelete }) => {
