+++ src/components/Invoices/InvoiceDetail.js
@@ -1,4 +1,5 @@
 import React from 'react';
+
 import { InvoiceDetailContainer, Section, SectionItems, SectionItemHeader, SectionItem , 
     SectionFooter }  from './InvoiceDetailStyle';
